+++ gluecodium/src/test/resources/smoke/dates/output/dart/lib/src/smoke/date_defaults.dart
@@ -5,14 +5,15 @@ class DateDefaults {
   DateTime dateTime;
   DateTime dateTimeUtc;
   DateTime beforeEpoch;
-  DateDefaults(this.dateTime, this.dateTimeUtc, this.beforeEpoch);
+  DateTime exactlyEpoch;
+  DateDefaults(this.dateTime, this.dateTimeUtc, this.beforeEpoch, this.exactlyEpoch);
   DateDefaults.withDefaults()
-    : dateTime = DateTime.fromMillisecondsSinceEpoch(1643966117000), dateTimeUtc = DateTime.fromMillisecondsSinceEpoch(1643966117000), beforeEpoch = DateTime.fromMillisecondsSinceEpoch(-1511793883000);
+    : dateTime = DateTime.fromMillisecondsSinceEpoch(1643966117000), dateTimeUtc = DateTime.fromMillisecondsSinceEpoch(1643966117000), beforeEpoch = DateTime.fromMillisecondsSinceEpoch(-1511793883000), exactlyEpoch = DateTime.fromMillisecondsSinceEpoch(0);
 }
 // DateDefaults "private" section, not exported.
 final _smokeDatedefaultsCreateHandle = __lib.catchArgumentError(() => __lib.nativeLibrary.lookupFunction<
-    Pointer<Void> Function(Uint64, Uint64, Uint64),
-    Pointer<Void> Function(int, int, int)
+    Pointer<Void> Function(Uint64, Uint64, Uint64, Uint64),
+    Pointer<Void> Function(int, int, int, int)
   >('library_smoke_DateDefaults_create_handle'));
 final _smokeDatedefaultsReleaseHandle = __lib.catchArgumentError(() => __lib.nativeLibrary.lookupFunction<
     Void Function(Pointer<Void>),
@@ -30,30 +31,39 @@ final _smokeDatedefaultsGetFieldbeforeEpoch = __lib.catchArgumentError(() => __l
     Uint64 Function(Pointer<Void>),
     int Function(Pointer<Void>)
   >('library_smoke_DateDefaults_get_field_beforeEpoch'));
+final _smokeDatedefaultsGetFieldexactlyEpoch = __lib.catchArgumentError(() => __lib.nativeLibrary.lookupFunction<
+    Uint64 Function(Pointer<Void>),
+    int Function(Pointer<Void>)
+  >('library_smoke_DateDefaults_get_field_exactlyEpoch'));
 Pointer<Void> smokeDatedefaultsToFfi(DateDefaults value) {
   final _dateTimeHandle = dateToFfi(value.dateTime);
   final _dateTimeUtcHandle = dateToFfi(value.dateTimeUtc);
   final _beforeEpochHandle = dateToFfi(value.beforeEpoch);
-  final _result = _smokeDatedefaultsCreateHandle(_dateTimeHandle, _dateTimeUtcHandle, _beforeEpochHandle);
+  final _exactlyEpochHandle = dateToFfi(value.exactlyEpoch);
+  final _result = _smokeDatedefaultsCreateHandle(_dateTimeHandle, _dateTimeUtcHandle, _beforeEpochHandle, _exactlyEpochHandle);
   dateReleaseFfiHandle(_dateTimeHandle);
   dateReleaseFfiHandle(_dateTimeUtcHandle);
   dateReleaseFfiHandle(_beforeEpochHandle);
+  dateReleaseFfiHandle(_exactlyEpochHandle);
   return _result;
 }
 DateDefaults smokeDatedefaultsFromFfi(Pointer<Void> handle) {
   final _dateTimeHandle = _smokeDatedefaultsGetFielddateTime(handle);
   final _dateTimeUtcHandle = _smokeDatedefaultsGetFielddateTimeUtc(handle);
   final _beforeEpochHandle = _smokeDatedefaultsGetFieldbeforeEpoch(handle);
+  final _exactlyEpochHandle = _smokeDatedefaultsGetFieldexactlyEpoch(handle);
   try {
     return DateDefaults(
       dateFromFfi(_dateTimeHandle),
       dateFromFfi(_dateTimeUtcHandle),
-      dateFromFfi(_beforeEpochHandle)
+      dateFromFfi(_beforeEpochHandle),
+      dateFromFfi(_exactlyEpochHandle)
     );
   } finally {
     dateReleaseFfiHandle(_dateTimeHandle);
     dateReleaseFfiHandle(_dateTimeUtcHandle);
     dateReleaseFfiHandle(_beforeEpochHandle);
+    dateReleaseFfiHandle(_exactlyEpochHandle);
   }
 }
 void smokeDatedefaultsReleaseFfiHandle(Pointer<Void> handle) => _smokeDatedefaultsReleaseHandle(handle);